Implement gather-temporaries transform and display helpers
It's nice to be able to see the shape of the transformed program, so
I've added ->sexpr (make a program into a s-expression close to what
the code would look like) and show (make a value into an s-expression
for display). The latter understands classes to some degree, unlike
the native SISC display.