Better error reporting in smoketests
Isolate each test so failures don't stop it proceeding.

I tried to make it show more readable errors ("no applicable method
for args <pages of glyphs>") using `show`, but it's all encoded into a
string by the time we get to handle it. A side effect of this attempt
was to put `->sexpr` in its place, apart from `show`, which is tidier:
->sexpr relies on the various classes being defined, whereas `show` is
general.