Mark the predef env to get the correct cons
When preparing or compiling a left-left-lambda with a dotted argument,
code is generated to cons the remainder arguments into a list. This
relies on finding the predefined procedure 'cons'.

Because there can now be different predefined environments (one for
compiled code, one for interpreted macro code), I now have to look for
the predefined environment rather than having it hard-wired. I'm just
marking the predef environment in the same way the global environment
is marked.